Fixes #137 Fixes #138
* Ticket refactoring - splitting up in separate files
* Created shouldn't be nil
* Fixing unit tests
